Rename RLS AI assistant to just PolicyEditorPanel and remove all RLS assistant stuff
