+++ apps/studio/components/interfaces/Auth/Policies/PolicyEditorPanel/LockedQuerySection.tsx
